+++ stellar-live-source-datalake/go.mod
@@ -1,12 +1,9 @@
-module github.com/stellar/ttp-processor-demo/stellar-live-source-datalake
+module github.com/withObsrvr/ttp-processor-demo/stellar-live-source-datalake
 
 go 1.21
 
 require (
-	github.com/stellar/go v0.0.0-20240301000000-000000000000
-	github.com/withObsrvr/stellar-cdp v0.0.0-20240301000000-000000000000
-	github.com/withObsrvr/stellar-datastore v0.0.0-20240301000000-000000000000
-	github.com/withObsrvr/stellar-ledgerbackend v0.0.0-20240301000000-000000000000
-	google.golang.org/grpc v1.61.0
-	google.golang.org/protobuf v1.32.0
+	github.com/stellar/go v0.0.0-20250322075617-5ccaef04a539
+	google.golang.org/grpc v1.71.0
+	google.golang.org/protobuf v1.36.5
 )
